+++ crates/api-desc/test.sh
@@ -13,9 +13,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-set -ex
+set -e
 
-cargo check
+. "$(git rev-parse --show-toplevel)"/scripts/test-helper.sh
+
+test_helper
+
 cargo test
-cargo fmt -- --check
-cargo clippy -- --deny=warnings