+++ crates/runner-nordic/src/main.rs
@@ -14,6 +14,7 @@
 
 #![no_std]
 #![no_main]
+#![feature(never_type)]
 #![feature(try_blocks)]
 
 extern crate alloc;
@@ -130,7 +131,9 @@ fn main() -> ! {
         .build();
     let rng = Rng::new(p.RNG);
     let ccm = Ccm::init(p.CCM, p.AAR, DataRate::_1Mbit);
-    let storage = Some(Storage::new(p.NVMC));
+    storage::init(p.NVMC);
+    let storage = Some(Storage::new_store());
+    crate::tasks::platform::update::init(Storage::new_other());
     let pins = uarte::Pins {
         txd: port0.p0_06.into_push_pull_output(gpio::Level::High).degrade(),
         rxd: port0.p0_08.into_floating_input().degrade(),