+++ crates/scheduler/src/call.rs
@@ -22,6 +22,7 @@ mod clock;
 mod crypto;
 mod debug;
 mod led;
+mod platform;
 mod rng;
 mod scheduling;
 mod store;
@@ -35,6 +36,7 @@ pub fn process<B: Board>(call: Api<DispatchSchedulerCall<B>>) {
         Api::Crypto(call) => crypto::process(call),
         Api::Debug(call) => debug::process(call),
         Api::Led(call) => led::process(call),
+        Api::Platform(call) => platform::process(call),
         Api::Rng(call) => rng::process(call),
         Api::Scheduling(call) => scheduling::process(call),
         Api::Store(call) => store::process(call),