+++ crates/store/fuzz/test.sh
@@ -13,9 +13,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-set -ex
+set -e
+
+. "$(git rev-parse --show-toplevel)"/scripts/test-helper.sh
+
+test_helper
 
 cargo check
-cargo fmt -- --check
-cargo clippy -- --deny=warnings
 cargo test --lib --examples